collector_probe: handle catch-all vs. direct syscall tracepoints
The Falco libs BPF driver is designed to catch all syscalls through
the sys_enter and sys_exit tracepoints.  The collector custom
probe attaches to particular system calls only.  In that case,
the tracepoint context depends on the syscall with a syscall-specific
format (structure).

This commit converts the syscall-specific format into the format
sys enter format expected by Falco's BPF driver and filler programs.
Previously, the custom probe copied data from beyond the tracepoint
context. This resulted in offset check failure and a permission denied
when attaching BPF probe (see also notes in the custom probe).

This is the sys enter path.  For the sys exit path, the syscall-specific
exit format (which is identical for direct attached syscalls) needs
to be aligned (wrt. to offsets) with the 'catch-all' sys exit format.

Signed-off-by: Hendrik Brueckner <brueckner@linux.ibm.com>